fix(api) : deployment connection folder refactored
