+++ packages/api/src/app.module.ts
@@ -6,8 +6,8 @@ import { ApikeyModule } from './server/api-key/service/apikey.module';
 import { ApplicationController } from './server/application/application.controller';
 import { ApplicationModule } from './server/application/service/application.module';
 import { AuthModule } from './server/auth/auth.module';
-import { DeploymentConnectionController } from './server/deployment-conncetion/deployment-connection.controller';
-import { DeploymentConnectionModule } from './server/deployment-conncetion/service/deployment-connection.module';
+import { DeploymentConnectionController } from './server/deployment-conncection/deployment-connection.controller';
+import { DeploymentConnectionModule } from './server/deployment-conncection/service/deployment-connection.module';
 import { ExceptionsModule } from './server/exceptions/exceptions.module';
 import { PropertyController } from './server/property/property.controller';
 import { PropertyModule } from './server/property/service/property.module';
