+++ packages/api/src/repository/data-services.abstract.ts
@@ -1,7 +1,7 @@
 import { ActivityStream } from 'src/server/analytics/activity-stream.entity';
 import { Apikey } from '../server/api-key/apikey.entity';
 import { Application } from '../server/application/application.entity';
-import { DeploymentConnection } from '../server/deployment-conncetion/deployment-connection.entity';
+import { DeploymentConnection } from '../server/deployment-conncection/deployment-connection.entity';
 import { Environment } from '../server/environment/environment.entity';
 import { EventTimeTrace } from '../server/event/event-time-trace.entity';
 import { Event } from '../server/event/event.entity';
