+++ packages/api/src/server/deployment-conncection/service/deployment-connection.factory.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { DeploymentConnection } from 'src/repository/mongo/model';
-import { DeploymentConnectionDTO, UpdateDeploymentConnectionDTO } from 'src/server/deployment-conncetion/deployment-connection.dto';
+import { DeploymentConnectionDTO, UpdateDeploymentConnectionDTO } from 'src/server/deployment-conncection/deployment-connection.dto';
 
 @Injectable()
 export class DeploymentConnectionFactoryService {
