+++ packages/api/src/server/deployment-conncection/service/deployment-connection.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { IDataServices } from 'src/repository/data-services.abstract';
 import { DeploymentConnection } from 'src/repository/mongo/model';
-import { DeploymentConnectionDTO, UpdateDeploymentConnectionDTO } from 'src/server/deployment-conncetion/deployment-connection.dto';
+import { DeploymentConnectionDTO, UpdateDeploymentConnectionDTO } from 'src/server/deployment-conncection/deployment-connection.dto';
 import { DeploymentConnectionFactoryService } from './deployment-connection.factory';
 
 @Injectable()
